refactor: modularize filter structure for JSR publication and type reusability
- Restructure from single file to directory structure for JSR compatibility:
  - main.ts: Entry point for ddu.vim (utilizing ddu-vim@11's directory support)
  - matcher.ts: Filter implementation (renamed from matcher_kensaku.ts)
  - types.ts: Type definitions for external consumption
- Add JSR package configuration (@milly/ddu-filter-matcher_kensaku)
- Add GitHub Actions workflow for automated JSR publishing
- Update CI/CD to include comprehensive quality checks
- Update README with type-safe configuration example using JSR imports
- Add package-specific LICENSE and README for JSR

This change enables type reusability through JSR publication, allowing
other plugins to import and use MatcherKensakuParams type definitions
directly from the registry.
