+++ denops/@ddu-filters/matcher_kensaku/LICENSE
@@ -1,0 +1,1 @@
+../../../LICENSE
