+++ denops/@ddu-filters/matcher_kensaku/README.md
@@ -1,0 +1,13 @@
+# ddu-filter-matcher_kensaku
+
+[![license:MIT](https://img.shields.io/github/license/Milly/ddu-filter-kensaku?style=flat-square)](LICENSE)
+
+Migemo matcher filter for [ddu.vim]
+
+This library provides a filter class and configuration types in a reusable form.
+
+If you want to use it as a Vim plugin, please refer to the [ddu-filter-kensaku]
+repository.
+
+[ddu.vim]: https://github.com/Shougo/ddu.vim
+[ddu-filter-kensaku]: https://github.com/Milly/ddu-filter-kensaku
