+++ denops/@ddu-filters/matcher_kensaku/types.ts
